requirements.txt
accelerate==1.2.0
peft==0.14.0
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
transformers==4.47.0
tokenizers==0.21.0
safetensors==0.4.5
huggingface-hub==0.26.5
filelock==3.9.0
PyYAML==6.0.1
Pillow==9.3.0
numpy
tqdm==4.65.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-image==0.23.2
scipy==1.11.4
scikit-learn==1.3.2
openai==1.105.0
requests
quadprog==0.1.13
tensorboard==2.20.0
wandb==0.21.1
pytorchcv==0.0.73
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
fsspec==2023.12.1
ftfy==6.2.0
importlib_metadata==8.5.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
networkx==3.0
packaging==23.1
pluggy==1.0.0
portalocker==3.0.0
protobuf==5.26.1
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.1
python-dateutil==2.8.2
sentencepiece==0.2.0
threadpoolctl==3.2.0
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.21.0
